backend/hybrid-detection/requirements.txt
# File: backend/hybrid-detection/requirements.txt
# Fungsi: Daftar semua library Python yang diperlukan untuk sistem hybrid

# Framework web untuk API
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Computer vision dan image processing
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1

# YOLO AI model
ultralytics==8.0.196
torch>=1.13.0
torchvision>=0.14.0

# Utilities
python-multipart==0.0.6
pydantic==2.4.2
requests==2.31.0

# Testing (opsional)
pytest==7.4.3